+++ src/ECS/Systems/LightingDirectional_S.h
@@ -51,15 +51,15 @@ public:
 		m_shapeQuad = Asset_Primitive::Create(engine, "quad");
 
 		// Preference Callbacks
-		m_renderSize.x = m_engine->addPrefCallback(PreferenceState::C_WINDOW_WIDTH, this, [&](const float &f) {
+		m_renderSize.x = (int)m_engine->addPrefCallback(PreferenceState::C_WINDOW_WIDTH, this, [&](const float &f) {
 			m_renderSize = glm::ivec2(f, m_renderSize.y);
 		});
-		m_renderSize.y = m_engine->addPrefCallback(PreferenceState::C_WINDOW_HEIGHT, this, [&](const float &f) {
+		m_renderSize.y = (int)m_engine->addPrefCallback(PreferenceState::C_WINDOW_HEIGHT, this, [&](const float &f) {
 			m_renderSize = glm::ivec2(m_renderSize.x, f);
 		});	
-		m_updateQuality = m_engine->addPrefCallback(PreferenceState::C_SHADOW_QUALITY, this, [&](const float &f) {m_updateQuality = f; });
-		m_shadowSize.x = m_engine->addPrefCallback(PreferenceState::C_SHADOW_SIZE_DIRECTIONAL, this, [&](const float &f) { m_shadowSize = glm::vec2(std::max(1.0f, f)); });
-		m_shadowSize = glm::vec2(std::max(1.0f, m_shadowSize.x));
+		m_updateQuality = (unsigned int)m_engine->addPrefCallback(PreferenceState::C_SHADOW_QUALITY, this, [&](const float &f) {m_updateQuality = (unsigned int)f; });
+		m_shadowSize.x = (int)m_engine->addPrefCallback(PreferenceState::C_SHADOW_SIZE_DIRECTIONAL, this, [&](const float &f) { m_shadowSize = glm::ivec2(std::max(1, (int)f)); });
+		m_shadowSize = glm::ivec2(std::max(1, m_shadowSize.x));
 		m_shadowFBO.resize(m_shadowSize, 4);
 		m_shader_Lighting->addCallback(this, [&](void) {m_shader_Lighting->setUniform(0, 1.0f / m_shadowSize.x); });
 
@@ -144,7 +144,7 @@ public:
 					float radius = glm::length(frustumCorners[7] - middle);
 					glm::vec3 aabb(radius);
 
-					const glm::vec3 volumeUnitSize = (aabb - -aabb) / m_shadowSize.x;
+					const glm::vec3 volumeUnitSize = (aabb - -aabb) / (float)m_shadowSize.x;
 					const glm::vec3 frustumpos = glm::vec3(shadowComponent->m_mMatrix * CamInv * glm::vec4(middle, 1.0f));
 					const glm::vec3 clampedPos = glm::floor((frustumpos + (volumeUnitSize / 2.0f)) / volumeUnitSize) * volumeUnitSize;
 					const glm::vec3 newMin = -aabb + clampedPos;
@@ -306,7 +306,7 @@ private:
 	GLuint m_quadVAO;
 	bool m_quadVAOLoaded;
 	unsigned int m_updateQuality;
-	glm::vec2 m_shadowSize;
+	glm::ivec2 m_shadowSize;
 	glm::ivec2	m_renderSize;
 	StaticBuffer m_indirectShape;
 	std::vector<GLuint> lightIndices, shadowIndices;
